Add option to mount EFS on $HOME
Previously, users had separate sessions across new EMR clusters.
With this change, every time a user logs in, they will have a
consistent file system, with the exact changes they saved last
time. Because we mount on $HOME, this includes configuration and
dot files.

Everything also works as before; i.e. if --email is not provided,
then the directory is not mounted, and Jupyer and Spark work as
intended.

The RRO file is downloaded in /mnt, rather than $HOME, to prevent
constant writing to EFS.

Bash config is written to /etc/profile.d/atmo.sh, which runs
whenever a use logs in.

Finally, this commit includes changes to the ansible deploy to
deploy an EFS volume.