Deploy Zeppelin notebook
Our current notebook story isn't great:
- opening a second notebook causes it to freeze (Bug 1290148);
- the Spark progress bar doesn't work reliably and can get stuck or
  disappear;
- Spark jobs can't be cancelled (Bug 1318706)
- jupyter can just freeze without returning any error whatsoever;
- Scala isn't supported (kernels for it do exist though)

The main issue with jupyter is that isn't wasn't build to work with an
external query engine. There is an effort underway to improve
the status quo [1] but it might take a while until it's production ready.

This patch introduces Zeppelin [2] to our stack which might solve some
of the above issues.  It provides features like multi-language support
within the same notebook [3], progress bars, job cancellation and an
interesting visualization story.

This is not a replacement for jupyter though, merely an experimental
tool that one day could graduate to a first-class citizen if it lives up
to its marketing claims.

[1] https://github.com/jupyter/jupyter/issues/212#issuecomment-271930788
[2] https://zeppelin.apache.org/
[3] https://www.zeppelinhub.com/viewer/notebooks/aHR0cHM6Ly9naXN0LmdpdGh1YnVzZXJjb250ZW50LmNvbS92aXRpbGxvL2JlODA0N2ZmZDg2MzM3ZDc0MzFhOGI3ZDI2NGUwMTNjL3Jhdy8zODEwNTE1ODM0OGVmYmYzYzFmMTYzZTA3NmNkMDJkZDczMTBmNTM5L0xhbmd1YWdlJTI1MjBtYWRuZXNzLmpzb24