Add option to mounts EFS on $HOME
This is the same commit as [0], with one difference. Previously,
dd was used to sudo append to the authorized_hosts file - but it
was not appending, instead it was overwriting.

This change uses tee, which does append to the file.

Bonus is that tee also echoes the string, so now the keys that
have been added can also be seen in the bootstrap log.

[0] 86407f1e968763fcbea3968447e22d30f996f111