Bug 1466936 - Install hyperloglog as a spark package
It looks like we were already pre-loading the vitillo version
of spark-hyperloglog from spark-packages.org for zeppelin,
but we've recently published the mozilla version of the package
on spark-packages.org; this PR makes that package a cluster default.

Packages specified in spark.jars.packages aren't installed at
cluster spinup time, but rather at instantiation of a SparkContext,
so this adds a slight amount of work for each new SparkContext
that's created.