Add pull request template
This is motivated by
[a discussion doc](https://docs.google.com/document/d/1NwY1iReuxcbvxCi9jRUOYiymCUOoSq3cnShCI-OiW3c/edit#heading=h.c6fp0v2ffacr)
that focused on the problem of implicit dependencies between repos.
emr-bootstrap-spark is one of the projects where changes can cause unanticipated
problems for running code that lives elsewhere, and it can be difficult to
track down the source of the problem when an emr-bootstrap-spark deploy goes
wrong.

This will be followed up by documentation changes to other repos that depend
on this one, making the dependency more explicit.